feat: createEvent 페이지 모바일 반응형 스타일링 + 인풋 유효성 검사 + 알림 모달 팝업
파일 업로드 관련된 것 외에는 모든 기능 구현 완료
