+++ client/src/components/molecule/createEvent/confirmbox.scss
@@ -1,3 +1,5 @@
+@import '../../../style/mediaQuery';
+
 .goback-modal-container {
   position: fixed;
   top: 0;
@@ -21,6 +23,7 @@
     pre {
       margin-top: 5rem;
       margin-bottom: 3.2rem;
+      line-height: 5.3rem;
     }
     .modal-foot {
       display: flex;
@@ -29,7 +32,7 @@
       button {
         all: unset;
         font-size: 1.5rem;
-        width: 18em;
+        width: 18rem;
         height: 6.5rem;
         font-weight: 600;
         &:nth-child(1) {
@@ -37,7 +40,11 @@
           color: #717174;
         }
         &:nth-child(2) {
-          border: 2px solid #ffffff;
+          border: 2px solid #fff;
+        }
+        &.one-btn {
+          border: 2px solid #fff;
+          color: #fff;
         }
         box-sizing: border-box;
         border-radius: 5rem;
@@ -45,6 +52,28 @@
         place-items: center;
         margin: 0 15px;
         cursor: pointer;
+      }
+    }
+  }
+}
+@include mobile {
+  .goback-modal-container {
+    padding: 0 3rem;
+    .inner-modal {
+      width: 100%;
+      height: 28%;
+      font-size: 1.7rem;
+      font-weight: 200;
+      pre {
+        line-height: 3rem;
+        margin: 1.5rem 0;
+        margin-top: 4rem;
+      }
+      .modal-foot {
+        button {
+          width: 14rem;
+          height: 5rem;
+        }
       }
     }
   }
