+++ server/package.json
@@ -25,6 +25,7 @@
     "moment": "^2.27.0",
     "mongoose": "^5.10.0",
     "morgan": "^1.10.0",
+    "multer": "^1.4.2",
     "path": "^0.12.7"
   },
   "devDependencies": {
